Move WeightDecayedAdam x10 definition into a separate file
